Swarm working and some progress on documentation and rationalisation of modules into package setup.
