+++ node.py
@@ -1,8 +1,64 @@
+"""
+Node module implements the core classes of MetaChem nodes. This includes the top level split between containers and
+control as well as the specific classes of nodes used in implementing MetaChem graphs.
+
+Classes
+-------
+ContainerNode
+    Super class  for container nodes in MetaChem
+
+ControlNode
+    Super clas for control nodes in MetaChem
+
+Termination
+    Node which indicated to the graph handler to stop running transitioning between nodes. This node has no outgoing
+    control edge.
+
+Action
+    Node which modifies particles in a sample. It has a single outgoing control edge.
+
+Decision
+    Node which branches control in the system. It has multiple outgoing control edges.
+
+Sampler
+    Node which moves particles between tanks and samples.
+
+Observer
+    Node which observes particles and environment and outputs information, sometimes summary statistics to the
+    environment.
+
+Tank
+    Container node which holds particles which can not be edited.
+
+Sample
+    Container node which holds particles which can be edited.
+
+Environment
+    Container node which holds non-particle system and environment information.
+"""
 import abc
 import random
 
 
 class ContainerNode(object):
+    """
+    A class for the basic description of a container node.
+
+    ...
+
+    Attributes
+    ----------
+    id: unique number to identify container node
+
+    Methods
+    -------
+    read()
+        Required to be implemented in subclasses
+    add(particles=None)
+        Required to be implemented in subclasses
+    remove(particles=None)
+        Required to be implemented in subclasses
+    """
     __metaclass__ = abc.ABCMeta
 
     @abc.abstractmethod
@@ -24,6 +80,32 @@ class ContainerNode(object):
 
 
 class ControlNode(object):
+    """
+    A class for the basic description of a control node.
+
+    ...
+
+    Attributes
+    ----------
+    id: unique number to identify container node
+
+    Methods
+    -------
+    transition()
+        Used by graph to perform the algorithmic action of the node. It runs the transition functions including checking
+        return from check function against a random number to allow for probabilistic execution of a nodes process.
+    read()
+        Required to be implemented in subclasses
+    check()
+        Required to be implemented in subclasses
+    pull()
+        Required to be implemented in subclasses
+    process()
+        Required to be implemented in subclasses
+    push()
+        Required to be implemented in subclasses
+
+    """
     __metaclass__ = abc.ABCMeta
 
     @abc.abstractmethod
@@ -61,6 +143,25 @@ class ControlNode(object):
 
 
 class Termination(ControlNode):
+    """
+    A class for the Termination node which indicates to the graph handler to stop processing the graph.
+
+    ...
+
+    Methods
+    -------
+    read()
+        Does nothing.
+    check()
+        Does nothing.
+    pull()
+        Does nothing.
+    process()
+        Does nothing.
+    push()
+        Does nothing.
+
+    """
 
     def __init__(self):
         super(Termination, self).__init__()
@@ -82,6 +183,32 @@ class Termination(ControlNode):
 
 
 class Action(ControlNode):
+    """
+    A class for the basic description of a control node.
+
+    ...
+
+    Attributes
+    ----------
+    writesample:    The sample node to which the node will push modified or new particles to. Must be a Sample.
+    readsample:     The sample node from which the node will pull particles to modify or use in forming new particles. Must
+                    be a Sample.
+    readcontainers: Any other particle or environment containers the node needs to read from for it's process.
+
+    Methods
+    -------
+    read()
+        Required to be implemented in subclasses
+    check()
+        Required to be implemented in subclasses. Defaults to returning 0 so defaults to process happening.
+    pull()
+        Required to be implemented in subclasses
+    process()
+        Required to be implemented in subclasses
+    push()
+        Required to be implemented in subclasses
+
+    """
     __metaclass__ = abc.ABCMeta
 
     @abc.abstractmethod
@@ -127,6 +254,34 @@ class Action(ControlNode):
 
 
 class Decision(ControlNode):
+    """
+    A class for the basic Decision node. This node it used to branch control function. It does this by returning a
+    number which is used to tell the graph handler which node to transition to next.
+
+    ...
+
+    Attributes
+    ----------
+    options: Integer indicating the number of outgoing graph edges.
+    readcontainers: List of containers which are read to provide information to base decision on.
+
+    Methods
+    -------
+    read()
+        Required to be implemented in subclasses.
+    check()
+        Uses super function.
+    pull()
+        Does nothing.
+    process()
+        Required to be implemented in subclasses. Returns an integer indication which control edge to follow to next
+        node. Defaults to returning 0.
+    push()
+        Does nothing.
+    transition()
+        Over writes normal transition function to return decision.
+
+    """
     __metaclass__ = abc.ABCMeta
 
     @abc.abstractmethod
@@ -161,6 +316,32 @@ class Decision(ControlNode):
 
 
 class Sampler(ControlNode):
+    """
+    A class for the basic description of a control node.
+
+    ...
+
+    Attributes
+    ----------
+    containersin: Containers from which the node can read and remove particles
+    containersout: Containers from which the node can read and add particles
+    readcontainers: Containers from which the node can read information
+    sample: List for storing particles in while moving them between containers
+
+    Methods
+    -------
+    read()
+        Required to be implemented in subclasses.
+    check()
+        Defaults to return 0.
+    pull()
+        Required to be implemented in subclasses.
+    process()
+        Does nothing.
+    push()
+        Required to be implemented in subclasses.
+
+    """
     __metaclass__ = abc.ABCMeta
 
     @abc.abstractmethod
@@ -197,6 +378,33 @@ class Sampler(ControlNode):
 
 
 class Observer(ControlNode):
+    """
+    A class for the basic description of an observer node. These are used to update variables and generate summary
+    statistics.
+
+    ...
+
+    Attributes
+    ----------
+    containersin: Containers from which the node can read and remove.
+    containersout: Containers from which the node can read and add.
+    readcontainers: Containers from which the node can read only.
+    index: Used to indicate the need to read a particular variable in a list.
+
+    Methods
+    -------
+    read()
+        Required to be implemented in subclasses
+    check()
+        Returns 0 using default behavior.
+    pull()
+        Required to be implemented in subclasses
+    process()
+        Required to be implemented in subclasses
+    push()
+        Required to be implemented in subclasses
+
+    """
     __metaclass__ = abc.ABCMeta
 
     @abc.abstractmethod
@@ -241,6 +449,21 @@ class Observer(ControlNode):
 
 
 class Tank(ContainerNode):
+    """
+    A class for the basic description of a Tank particle container node. This one can not be connected as a writeable
+    containers to an action.
+
+    ...
+
+    Methods
+    -------
+    read()
+        Required to be implemented in subclasses
+    add(particles=None)
+        Required to be implemented in subclasses
+    remove(particles=None)
+        Required to be implemented in subclasses
+    """
     __metaclass__ = abc.ABCMeta
 
     @abc.abstractmethod
@@ -262,6 +485,22 @@ class Tank(ContainerNode):
 
 
 class Sample(ContainerNode):
+    """
+    A class for the basic description of a Sample particle container node. This one can be connected as a writeable
+    containers to an action.
+
+    ...
+
+    Methods
+    -------
+    read()
+        Required to be implemented in subclasses
+    add(particles=None)
+        Required to be implemented in subclasses
+    remove(particles=None)
+        Required to be implemented in subclasses
+    """
+
     __metaclass__ = abc.ABCMeta
 
     @abc.abstractmethod
@@ -283,6 +522,20 @@ class Sample(ContainerNode):
 
 
 class Environment(ContainerNode):
+    """
+    A class for the basic description of an environment node. This can container any information needed by the system.
+    ...
+
+    Methods
+    -------
+    read()
+        Required to be implemented in subclasses
+    add(particles=None)
+        Required to be implemented in subclasses
+    remove(particles=None)
+        Required to be implemented in subclasses
+    """
+
     __metaclass__ = abc.ABCMeta
 
     def __init__(self):
